Position: handle when application state skips inactive
Application state can go directly from Active to (say) Suspended. The
old code didn't anticipate this and would fail to stop using position
in this case.

Instead of explicitly check for Active state, this commit decides that
the app is going inactive if the state is anything but Active.
m_applicationActive variable has already been used to prevent double
handling, making it safe to do so.

Fixes: https://github.com/ubports/ubuntu-touch/issues/1226